renamed and finished decrypter
